web_hws/HW3/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
requests
python-dotenv
pydantic
requests